+++ CHANGELOG.md
@@ -3,3 +3,15 @@
 ## [0.0.3] - 2024-10-01
 ### Added
 - Initial release of the NestJS logging module.
+
+## [0.0.4] - 2024-10-02
+### Added
+- Added the LogExporter option to export logs based on the parameter: file, console, or both.
+
+## [0.0.6] - 2024-10-02
+### Added
+- Pino logger file export issue solved
+
+## [0.0.7] - 2024-10-02
+### Added
+- Update package description